Move sheets abstraction to its owj package
