+++ collect_app/src/main/java/org/odk/collect/android/gdrive/GoogleDriveApi.java
@@ -5,6 +5,8 @@ import com.google.api.services.drive.Drive;
 import com.google.api.services.drive.model.FileList;
 import com.google.api.services.drive.model.Permission;
 
+import org.odk.collect.android.gdrive.sheets.DriveApi;
+
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
