+++ collect_app/src/main/java/org/odk/collect/android/gdrive/InstanceGoogleSheetsUploader.java
@@ -38,6 +38,10 @@ import org.odk.collect.android.dao.FormsDao;
 import org.odk.collect.android.exception.BadUrlException;
 import org.odk.collect.android.exception.MultipleFoldersFoundException;
 import org.odk.collect.android.forms.Form;
+import org.odk.collect.android.gdrive.sheets.DriveApi;
+import org.odk.collect.android.gdrive.sheets.DriveHelper;
+import org.odk.collect.android.gdrive.sheets.SheetsApi;
+import org.odk.collect.android.gdrive.sheets.SheetsHelper;
 import org.odk.collect.android.instances.Instance;
 import org.odk.collect.android.preferences.GeneralKeys;
 import org.odk.collect.android.preferences.GeneralSharedPreferences;
