+++ collect_app/src/main/java/org/odk/collect/android/gdrive/sheets/DriveHelper.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package org.odk.collect.android.gdrive;
+package org.odk.collect.android.gdrive.sheets;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
